requirements.txt
firebase_admin==6.5.0
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
Pillow==10.3.0
plotly==5.22.0
Requests==2.32.2
scikit_learn==1.3.2
streamlit==1.34.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
